requirements.txt
langchain
langchain-community
python-dotenv
streamlit
langchain-experimental
sentence-transformers
langchain-chroma
langchainhub
unstructured
langchain-google-genai

